Optimized simple agro check
Due to the fact that the game is single player and there is only one
enemy AI agent, there is no sense in performing a Multi Sphere Trace.
Instead, it is enough to check the distance of the player to the agent,
and then perform the visibility check by just one line trace.