Fixed Eye movement outside pyramid
The Eye is now properly blocked in the entrance of the pyramid, using a
Blocking Volume that only blocks the newly added object type ‘AI’
(which is the same as ‘Pawn’ for all other purposes).
The Eye keeps on chasing the player but doesn’t step out of the pyramid.
This fixes the previous behaviour, where the Eye would randomly lose
vision of the player when stepped out of the pyramid, and sometimes
even successfully follow them outside of the NavMesh, with
unpredictable results.