Eye balance tuning
Tuned up eye movement speed and vision radius.
Fixed NavMesh settings. The eye is too big for the navmesh to generate
in certain corridors. However, the eye is already quite small. Rather
than making it even smaller, the capsule is reduced to allow navigation
through all corridors, but the actual mesh is kept the same size with
ignored collisions. This allow the eye to stay at a reasonable size
while functioning properly with the navmesh, at the expense of going
through some walls and corners (the frequency of this, as well as the
actual overlapping, are low enough to justify the workaround).